Update CeloDao.sol
Errors:

In the createProposal function, the amount parameter is not checked for overflow. This could allow a malicious user to create a proposal for an extremely large amount of money, which could overflow the contract's balance and cause the contract to malfunction.

In the performVote function, the proposalId parameter is not checked for validity. This could allow a malicious user to vote on a proposal that does not exist, which could lead to unexpected behavior.

In the payBeneficiary function, the proposalId parameter is not checked for validity. This could allow a malicious user to call the function with an invalid proposal ID, which could lead to unexpected behavior.

Typos:

In the stakeholderOnly modifier, the message parameter is spelled incorrectly. It should be "Only stakeholders are allowed to create Proposals" instead of "Only stakeholders are allowed to creat Proposals".

In the contributorOnly modifier, the message parameter is spelled incorrectly. It should be "Only contributors can make payment" instead of "Only contributors can make paymet".

Bugs:

The handleVoting function does not properly check for double voting. A malicious user could vote on a proposal multiple times, which could skew the vote results.
Possible fixes:

To fix the overflow issue in the createProposal function, you can check the amount parameter against the contract's balance before creating the proposal. If the amount parameter is greater than the contract's balance, you can revert the transaction.

To fix the validity issue in the performVote function, you can check the proposalId parameter against the totalProposals variable. If the proposalId parameter is greater than or equal to totalProposals, you can revert the transaction.

To fix the validity issue in the payBeneficiary function, you can check the proposalId parameter against the raisedProposals mapping. If the proposalId parameter is not in the raisedProposals mapping, you can revert the transaction.

To fix the typo in the stakeholderOnly modifier, you can change the message parameter to "Only stakeholders are allowed to create Proposals".

To fix the typo in the contributorOnly modifier, you can change the message parameter to "Only contributors can make payment".

To fix the double voting issue in the handleVoting function, you can maintain a set of voted-on proposals for each stakeholder. When a stakeholder tries to vote on a proposal, you can check if the stakeholder has already voted on that proposal. If so, you can revert the transaction.